requirements.txt
Django==4.0.4
django-extensions==3.1.5
psycopg2==2.9.3
django-import-export==2.8.0
